Import SimpleWebRTC 2.2.3 and other dependencies to be customized
SimpleWebRTC 2.2.3 is the same version provided by the current
"js/simplewebrtc.js" file. This file will be eventually removed and
replaced by a custom bundle of the imported files and its dependencies.

The dependencies are mostly defined in "package.json" but, just like
SimpleWebRTC itself, some of the dependencies will need to be
customized; the dependencies to be customized are getscreenmedia 5.1.2
and localmedia 5.2.0. Note that the imported SimpleWebRTC has other
dependencies too, but their uses will be removed from the code and thus
are not needed.

The relevant files were imported into "js/simplewebrtc/".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>